18/9 Evening
This evening I finished the searching function of the IsFlying parameter, so now when the flightless button is pushed it displays a weka, but when the flying button is pushed it displays a tui. This is all modular so I can continue adding more types of parameters with ease. Next I have to implement the colour searching function.