FINAL ASSESSMENT COMMIT
These are my final changes to the app before the assessment is due. Today I decided not to add any more features, as the two that are implemented already work well enough as it is, and the hassle of coding in new ones so close to the deadline is something I don't want to deal with. I fixed a bug where the search results' background would have a cyan splotch, added some more comments to less readable code, and removed a few scripts that I was no longer using.